create a constants module, extract constants from utility and main to new module
